+++ src/info.py
@@ -2,5 +2,5 @@ class Info:
     Author = 'Andreas Pfau'
     Domain = 'apfau.de'
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.40b23"  # see https://peps.python.org/pep-0440/
-    AppDateStr = "2025-09-26"
+    AppVersionStr = "0.40b24"  # see https://peps.python.org/pep-0440/
+    AppDateStr = "2025-09-30"